AccompanyingPersons: hack for reception paid vs. free
The previous implementation stored reception in
    bool AccompanyingPersons.reception
, so we can't store {0, free, paid} in that field.

A better implementation might have stored that as a string, but I'm
reluctant to try changing the type of existing data.  To resolve this,
I'm adding a
    bool AccompanyingPersons.receptionPaid

New registrations will now set .receptionPaid rather than .reception;
and the backend only counts .receptionPaid as having a cost.